Workshop_4_Simulation/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
joblib>=1.2.0
pytest>=7.3.0
pyyaml>=6.0
jupyter>=1.0.0
